backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0
google-generativeai>=0.8.0
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
anthropic==0.8.1
mem0ai==0.1.14
together==0.2.7
python-socketio==5.10.0
eventlet==0.33.3
pydantic==2.5.3
cryptography>=41.0.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
psutil==5.9.8
tenacity==8.2.3
aiohttp==3.9.1
docker==6.1.3
# Gemini Live Studio dependencies
google-genai>=0.2.0
pyaudio>=0.2.11
numpy>=1.24.0
websockets>=11.0
asyncio-socketio>=0.6.0
# Enhanced functionality dependencies
PyGithub==2.4.0
pyyaml==6.0.2
beautifulsoup4==4.12.3
selenium==4.25.0
webdriver-manager==4.0.2
# Code-server specific dependencies
google-cloud-run==0.10.5
google-cloud-container==2.46.0
google-cloud-build==3.26.0